segmentation/requirements.txt
mmcv>=0.2.10
numpy
matplotlib
six
terminaltables
pycocotools
torch>=1.1
torchvision
imagecorruptions
jpeg2dct
PyTurboJPEG